Log errors in linters with ERROR logging level
Prior to this fix when linters fail due to errors, the error is logged
with INFO logging level. As a result, due to the default logging level
of WARNING, the errors do not appear to the user which may leave the
user confused about why the linter did not work.

Therefore, increase the logging level of linter errors to ERROR so that
if a linter fails, the error appears to the user on the console.